restruct insert nchw and nhwc convert op
