+++ oneflow/ir/include/OneFlow/OneFlowBase.td
@@ -39,6 +39,7 @@ def OneFlow_IsImportCompatible : NativeOpTrait<"IsImportCompatible">;
 def OneFlow_AlternativeOp : NativeOpTrait<"IsAlternative">;
 def OneFlow_TensorSource : NativeOpTrait<"TensorSource">;
 def OneFlow_OnlyExistsInIR : NativeOpTrait<"OnlyExistsInIR">;
+def OneFlow_ElementwiseOp : NativeOpTrait<"IsElementwise">;
 
 class OneFlow_IROp<string mnemonic, list<Trait> traits = []> :
         Op<OneFlow_Dialect, mnemonic, !listconcat(traits, [OneFlow_OnlyExistsInIR])> {}
