+++ oneflow/ir/include/OneFlow/OneFlowOpTraits.h
@@ -127,6 +127,9 @@ class TensorSource : public TraitBase<ConcreteType, TensorSource> {
 template<typename ConcreteType>
 class OnlyExistsInIR : public TraitBase<ConcreteType, OnlyExistsInIR> {};
 
+template<typename ConcreteType>
+class IsElementwise : public TraitBase<ConcreteType, IsElementwise> {};
+
 }  // namespace OpTrait
 
 }  // namespace mlir
